Value class abstraction
The Value class has been abstracted so that different bitwidths can use different logic more effectively:

-64 bit values and below use the new LongValue class which uses most logic from
-65 bits and above use the LongArrayValue class, which uses new logic to exclusively treat higher bitwidth logic.

The abstracted Value class still has every Create method which choose which subclass to use. The Value also contains its constants and the bitwidth of the value. Every method has been reorganized for a better separation.

The internal variables(value, error and unknown) are currently package private instead of just private, so it is possible to access them from the Value and its other subclasses. If this turns out to be problematic we may consider moving Value and its subclasses into their own package.

fromLogString and radixOfLogString have *temporarily* moved to TestVector, as its the only class that uses those methods and seems to use some old(and maybe redundant) logic, so it requires some investigation(together with Test Vector itself)
