+++ src/main/java/com/cburch/logisim/data/TestVector.java
@@ -173,7 +173,7 @@ public class TestVector {
           vals[pinIndex] = Value.createUnknown(columnWidth[pinIndex]);
         } else {
           try {
-            vals[pinIndex] = Value.fromLogString(columnWidth[pinIndex], t);
+            vals[pinIndex] = fromLogString(columnWidth[pinIndex], t);
           } catch (Exception e) {
             String errorMsg = e.getMessage();
             // Enhance error message with column name if it's a "too many bits" error
@@ -193,7 +193,7 @@ public class TestVector {
 
         if (TestVector.this.data.isEmpty()) {
           if (!dcFlags[pinIndex] && !floatFlags[pinIndex]) {
-            columnRadix[pinIndex] = Value.radixOfLogString(columnWidth[pinIndex], t);
+            columnRadix[pinIndex] = radixOfLogString(columnWidth[pinIndex], t);
           }
         }
         pinIndex++;
@@ -340,5 +340,134 @@ public class TestVector {
       }
     }
     return null;
+  }
+
+    /**
+   * Code taken from Cornell's version of Logisim: http://www.cs.cornell.edu/courses/cs3410/2015sp/
+   */
+  private static Value fromLogString(BitWidth width, String t) throws Exception {
+    // Strip underscores from the string for readability (e.g., 0x0000_1111 -> 0x00001111)
+    // This must be done before radix detection since radixOfLogString uses length
+    final var sb = new StringBuilder(t.length());
+    for (int i = 0; i < t.length(); i++) {
+      final var c = t.charAt(i);
+      if (c != '_') {
+        sb.append(c);
+      }
+    }
+    final var cleaned = sb.toString();
+
+    final var radix = radixOfLogString(width, cleaned);
+    int offset;
+
+    if (radix == 16 && cleaned.startsWith("0x")) offset = 2;
+    else if (radix == 8 && cleaned.startsWith("0o")) offset = 2;
+    else if (radix == 2 && cleaned.startsWith("0b")) offset = 2;
+    else if (radix == 10 && cleaned.startsWith("-")) offset = 1;
+    else offset = 0;
+
+    int n = cleaned.length();
+
+    if (n <= offset) throw new Exception("expected digits");
+
+    int w = width.getWidth();
+    long value = 0;
+    long unknown = 0;
+
+    for (var i = offset; i < n; i++) {
+      final var c = cleaned.charAt(i);
+      int d;
+
+      if (c == 'x' && radix != 10) d = -1;
+      else if ('0' <= c && c <= '9') d = c - '0';
+      else if ('a' <= c && c <= 'f') d = 0xa + (c - 'a');
+      else if ('A' <= c && c <= 'F') d = 0xA + (c - 'A');
+      else
+        throw new Exception(
+            "Unexpected character '" + cleaned.charAt(i) + "' in \"" + t + "\"");
+
+      if (d >= radix)
+        throw new Exception("Unexpected character '" + cleaned.charAt(i) + "' in \"" + t + "\"");
+
+      value *= radix;
+      unknown *= radix;
+
+      if (radix != 10) {
+        if (d == -1) unknown |= (radix - 1);
+        else value |= d;
+      } else {
+        if (d == -1) unknown += (radix - 1);
+        else value += d;
+      }
+    }
+    if (radix == 10 && cleaned.charAt(0) == '-') {
+      value = -value;
+    }
+
+    // Check bit width - for signed values, check the range instead of bit shift
+    if (w == 64) {
+      if (((value & 0x7FFFFFFFFFFFFFFFL) >> (w - 1)) != 0) {
+        int actualBits = 64 - Long.numberOfLeadingZeros(value & 0x7FFFFFFFFFFFFFFFL);
+        throw new Exception("Too many bits in \"" + t + "\" expected " + w + " bit" + (w != 1 ? "s" : "")
+            + (actualBits > 0 ? " did you mean [" + actualBits + "]?" : ""));
+      }
+    } else {
+      // For signed decimal, check if value fits in w-bit signed range
+      if (radix == 10) {
+        long maxPositive = (1L << (w - 1)) - 1;
+        long minNegative = -(1L << (w - 1));
+        if (value > maxPositive || value < minNegative) {
+          // Calculate actual bits needed (for absolute value)
+          long absValue = value < 0 ? -value : value;
+          int actualBits = absValue == 0 ? 1 : 64 - Long.numberOfLeadingZeros(absValue) + 1; // +1 for sign bit
+          throw new Exception("Too many bits in \"" + t + "\" expected " + w + " bit" + (w != 1 ? "s" : "")
+              + (actualBits > 0 ? " did you mean [" + actualBits + "]?" : ""));
+        }
+        // Mask to width for signed values (two's complement representation)
+        long mask = (1L << w) - 1;
+        value &= mask;
+      } else {
+        // For unsigned (hex, octal, binary), use bit shift check
+        if ((value >> w) != 0) {
+          // Calculate actual bits needed
+          int actualBits = value == 0 ? 1 : 64 - Long.numberOfLeadingZeros(value);
+          String reminder = "";
+
+          // For hex values, suggest based on number of hex digits * 4 (each hex digit = 4 bits)
+          if (radix == 16 && cleaned.length() > 2) {
+            int hexDigits = cleaned.length() - 2; // Subtract "0x" prefix
+            // Use hex digits * 4 as the suggested bit width (each hex digit = 4 bits)
+            actualBits = hexDigits * 4;
+            reminder = " Remember that 0x means hex and each hex digit is 4 bits";
+          } else if (radix == 2 && cleaned.length() > 0) {
+            // For binary values, suggest based on number of binary digits (each binary digit = 1 bit)
+            int binaryDigits = cleaned.length() - (cleaned.startsWith("0b") ? 2 : 0); // Subtract "0b" prefix if present
+            // Use binary digits as the suggested bit width (each binary digit = 1 bit)
+            actualBits = binaryDigits;
+            reminder = " Remember that 0b means binary and each binary digit is 1 bit";
+          } else if (radix == 8 && cleaned.length() > 2) {
+            // For octal values, suggest based on number of octal digits * 3 (each octal digit = 3 bits)
+            int octalDigits = cleaned.length() - 2; // Subtract "0o" prefix
+            // Use octal digits * 3 as the suggested bit width (each octal digit = 3 bits)
+            actualBits = octalDigits * 3;
+            reminder = " Remember that 0o means octal and each octal digit is 3 bits";
+          }
+
+          throw new Exception("Too many bits in \"" + t + "\" expected " + w + " bit" + (w != 1 ? "s" : "")
+              + (actualBits > 0 ? " did you mean [" + actualBits + "]?" : "") + reminder);
+        }
+      }
+    }
+
+    unknown &= ((1L << w) - 1);
+    return Value.create_unsafe(w, 0, unknown, value);
+  }
+
+  private static int radixOfLogString(BitWidth width, String t) {
+    if (t.startsWith("0x")) return 16;
+    if (t.startsWith("0o")) return 8;
+    if (t.startsWith("0b")) return 2;
+    if (t.length() == width.getWidth()) return 2;
+    return 10;
   }
 }
